% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDM_fun.R
\name{permanovaFL}
\alias{permanovaFL}
\title{PERMANOVA test of association}
\usage{
permanovaFL(formula, data = .GlobalEnv, tree = NULL,
  dist.method = "bray", cluster.id = NULL, strata = NULL,
  how = NULL, perm.within.type = "free", perm.between.type = "none",
  perm.within.ncol = 0, perm.within.nrow = 0, n.perm.max = 5000,
  n.rej.stop = 20, seed = NULL, square.dist = TRUE,
  center.dist = TRUE, scale.otu.table = TRUE)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted in the form
of \code{data.matrix ~ sets of covariates} or \code{data.matrix |
confounders ~ sets of covariates}. The details of model specification are
given in "Details" of \code{ldm}. Additionally, in \code{permanovaFL}, the \code{data.matrix}
can be either an OTU table or a distance matrix. If it is an OTU table,
the distance matrix will be calculated internally using the OTU table, \code{tree} (if required), and 
\code{dist.method}. If \code{data.matrix} is a distance
matrix (having class \code{dist} or \code{matrix}), it can be squared and//or centered by
specifying \code{square.dist} and \code{center.dist} (described below).  Distance matrices are distinguished
from OTU tables by checking for symmetry of \code{as.matrix(data.matrix)}.}

\item{data}{an optional data frame, list or environment (or object coercible 
to a dataframe) containing the covariates of interest and confounding covariates. 
If not found in \code{data}, the covariates are taken from environment(formula), 
typically the environment from which \code{permanovaFL} is called. The default is .GlobalEnv.}

\item{tree}{a phylogenetic tree. Only used for calculating a 
phylogenetic-tree-based distance matrix. Not needed if the calculation of 
the requested distance does not involve a phylogenetic tree, or if a 
distance matrix is directly imported through \code{formula}.}

\item{dist.method}{method for calculating the distance measure, partial
match to all methods supported by \code{vegdist} in the \code{vegan} package
 (i.e., "manhattan", "euclidean", "canberra", "bray", "kulczynski", "jaccard", "gower", 
 "altGower", "morisita", "horn", "mountford", "raup" , "binomial", "chao", "cao", "mahalanobis")
  as well as "hellinger" and "wt-unifrac". 
  Not used if a distance matrix is specified in \code{formula}.
  The default is "bray". 
  For more details, see the \code{dist.method} argument in the \code{ldm} function.}

\item{cluster.id}{cluster identifiers. The default is value of NULL should be used if the observations are 
not in clusters (i.e., independent).}

\item{strata}{a factor variable (or, character variable converted into a factor) to define strata (groups), within which to constrain permutations. 
The default is NULL.}

\item{how}{a permutation control list, for users who want to specify their permutation control list using the \code{how} function 
from the \code{permute} R package.  The default is NULL.}

\item{perm.within.type}{a character string that takes values "free", "none", "series", or "grid".
The default is "free" (for random permutations).}

\item{perm.between.type}{a character string that takes values "free", "none", or "series".
The default is "none".}

\item{perm.within.ncol}{a positive integer, only used if perm.within.type="grid". 
The default is 0.  See the documentation for the R package \code{permute} for further details.}

\item{perm.within.nrow}{a positive integer, only used if perm.within.type="grid". 
The default is 0.  See the documentation for the R package \code{permute} for further details.}

\item{n.perm.max}{the maximum number of permutations.
The default is 5000.}

\item{n.rej.stop}{the minimum number of rejections (i.e., the permutation 
statistic exceeds the observed statistic) to obtain before stopping. The 
default is 20.}

\item{seed}{an integer seed for the random number generator in the 
permutation procedure. The default is NULL; with the default value, an integer seed will be 
generated internally and randomly. In either case, the integer seed will be 
stored in the output object.}

\item{square.dist}{a logical variable indicating whether to square the
distance matrix. The default is TRUE.}

\item{center.dist}{a logical variable indicating whether to center the 
distance matrix as described by Gower (1966). The default is TRUE.}

\item{scale.otu.table}{a logical variable indicating whether to scale the OTU table.
For count data, this corresponds to dividing by the library size to give
relative frequencies.  The default is TRUE.}
}
\value{
a list consisting of 
\item{F.statistics}{F statistics for the global test of each set of covariates}
\item{p.permanova}{p-values for the global test
  of each set of covariates} 
\item{n.perm.completed}{number of permutations completed}
\item{permanova.stopped}{a logical value indicating whether the 
  stopping criterion has been met by all global tests}
\item{seed}{a single-value integer seed that is user supplied or internally generated}
}
\description{
This function performs the PERMANOVA test that can allow adjustment of
confounders and control of clustered data. As in \code{ldm}, 
\code{permanovaFL} allows multiple sets of covariates to be tested, 
in the way that the sets are entered sequentially and the variance 
explained by each set is that part that remains after the previous 
sets have been fit.
}
\examples{
res.permanova <- permanovaFL(formula=throat.otu.tab | (Sex+AntibioticUse) ~ SmokingStatus+PackYears, 
                            data=throat.meta, dist.method="bray", seed=123)
}
\author{
Yi-Juan Hu <yijuan.hu@emory.edu>, Glen A. Satten <gas0@cdc.gov>
}
\keyword{microbiome}
