% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.meta.R
\docType{data}
\name{sim.meta}
\alias{sim.meta}
\title{Metadata of the simulated microbiome samples}
\format{
A data frame with 100 observations on 5 variables.
}
\usage{
data("sim.meta")
}
\description{
This data set includes 100 simulated samples from 50 subjects, with two samples from each subject. 
The metadata contain subject IDs and data from four variables, two (X.between and Y.between) varying between subjects 
and two (X.within and Y.within) varying within subjects.
}
\examples{
data(sim.meta)
}
\keyword{datasets}
