% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.otu.tab.R
\docType{data}
\name{sim.otu.tab}
\alias{sim.otu.tab}
\title{OTU count table of the simulated microbiome samples}
\format{
A data frame with 100 observations on 813 variables.
}
\usage{
data("sim.otu.tab")
}
\description{
This is the OTU table corresponding to the 100 simulated samples in sim.meta.
The samples in the rows of the OTU table match with the samples in the rows of sim.meta.
}
\examples{
data(sim.otu.tab)
}
\keyword{datasets}
