% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.otu.tab5.R
\docType{data}
\name{sim.otu.tab5}
\alias{sim.otu.tab5}
\title{Filtered OTU count table of the simulated microbiome samples}
\format{
A data frame with 100 observations on 593 variables.
}
\usage{
data("sim.otu.tab5")
}
\description{
This table was derived from sim.otu.tab, after filtering out OTUs that are present (having non-zero counts) 
in less than five samples. This filter reduced the number of OTUs from 813 to 593.
}
\examples{
data(sim.otu.tab5)
}
\keyword{datasets}
