% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throat.otu.tab5.R
\docType{data}
\name{throat.otu.tab5}
\alias{throat.otu.tab5}
\title{Filtered OTU count table from 16S sequencing of the throat microbiome samples}
\format{
A data frame with 60 observations on 233 variables.
}
\usage{
data("throat.otu.tab5")
}
\description{
This table was derived from throat.otu.tab, after filtering out OTUs that are present (having non-zero counts) 
in less than five samples. This filter reduced the number of OTUs from 856 to 233.
}
\examples{
data(throat.otu.tab5)
}
\keyword{datasets}
