% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throat.tree.R
\docType{data}
\name{throat.tree}
\alias{throat.tree}
\title{UPGMA tree of the OTUs from 16S sequencing of the throat microbiome samples}
\format{
List of 4 data frames.
}
\source{
Charlson ES, Chen J, Custers-Allen R, Bittinger K, Li H, et al. (2010) 
Disordered Microbial Communities in the Upper Respiratory Tract of Cigarette Smokers. PLoS ONE 5(12): e15216.
}
\usage{
data("throat.tree")
}
\description{
The OTU tree is constructed using UPGMA on the K80 distance matrix of the OTUs. 
It is a rooted tree of class "phylo".
}
\examples{
data(throat.tree)
}
\references{
R package "GUniFrac"
}
\keyword{datasets}
